+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.14.0] - 2024-08-17
 
 ### Changed
 
@@ -336,6 +336,7 @@ No engines except [testkit] are able to provide a meaningful implementation of
 [0.12.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.12.1
 [0.13.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.13.0
 [0.13.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.13.1
+[0.14.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.14.0
 
 <!-- adr references -->